refactored + normed exec, need to fix fd open for: Makefile cat > out | ls -l | < infile cat > in1
